use 16-true precision

+++ songsuro/autoencoder/train.py
@@ -54,7 +54,7 @@ def train(
 		callbacks=[tqdm_cb, ckpt_cb, early_stop_callback],
 		check_val_every_n_epoch=1,
 		log_every_n_steps=1,
-		precision=16,
+		precision="16-true",
 		devices=2,
 		strategy=ModelParallelStrategy(),
 		num_sanity_val_steps=0,
